Finished plane task, not tested yet

+++ support/shared_mem.h
@@ -7,7 +7,7 @@
 
 #define SHARED_MEM_MAX_OBJ 8
 
-int shared_mem_create(int id, int size);
+int shared_mem_create(int id, size_t size);
 int shared_mem_update(int id, const void* pointer);
 void* shared_mem_get(int id);
 int shared_mem_delete(int id);
